+++ CHANGELOG.md
@@ -24,6 +24,7 @@ n/a
 * Added `androidx.core:core-splashscreen:1.0.1` library to enable animated splash screens and branding drawables.
 * Added animated splash screen.
 * Removed bottom navigation bar from `MainActivity`, since the home screen is completely empty and useless. The main activity now displays the list of entries and the settings can be opened by clicking a button within the app bar.
+* Viewing and editing entries now happens through the same activity.
 
 <br/>